+++ components/frontend/src/utils.js
@@ -56,7 +56,7 @@ export function capitalize(string) {
 
 export function pluralize(word, count) {
     // Pluralize (naively; it doesn't work for words like sheep) the word if count > 1
-    return word + (count <= 1 ? "" : "s");
+    return word + (count === 1 ? "" : "s");
 }
 
 export function nice_number(number) {